+++ examples/pubsub_realtime/nodeset/pubsub_nodeset_rt_subscriber.c
@@ -199,9 +199,9 @@ changePubSubApplicationCallback(UA_Server *server, UA_NodeId identifier, UA_UInt
 /* Remove the callback added for cyclic repetition */
 static void
 removePubSubApplicationCallback(UA_Server *server, UA_NodeId identifier, UA_UInt64 callbackId) {
-    if(callbackId && (pthread_join(callbackId, NULL) != 0))
+    if(callbackId && (pthread_join((long unsigned int)callbackId, NULL) != 0))
         UA_LOG_WARNING(UA_Log_Stdout, UA_LOGCATEGORY_USERLAND,
-                       "Pthread Join Failed thread: %ld\n", callbackId);
+                       "Pthread Join Failed thread: %lld\n", callbackId);
 }
 
 /**
@@ -670,7 +670,7 @@ int main(int argc, char **argv) {
         size_t subLoopVariable               = 0;
         for (subLoopVariable = 0; subLoopVariable < measurementsSubscriber;
              subLoopVariable++) {
-             fprintf(fpSubscriber, "%ld,%ld.%09ld,%lf\n",
+             fprintf(fpSubscriber, "%lld,%ld.%09ld,%lf\n",
                      subscribeCounterValue[subLoopVariable],
                      subscribeTimestamp[subLoopVariable].tv_sec,
                      subscribeTimestamp[subLoopVariable].tv_nsec,
@@ -682,7 +682,7 @@ int main(int argc, char **argv) {
         size_t subLoopVariable               = 0;
         for (subLoopVariable = 0; subLoopVariable < measurementsSubscriber;
              subLoopVariable++) {
-             fprintf(fpSubscriber, "%ld,%ld.%09ld,%lf\n",
+             fprintf(fpSubscriber, "%lld,%ld.%09ld,%lf\n",
                      subscribeCounterValue[subLoopVariable],
                      subscribeTimestamp[subLoopVariable].tv_sec,
                      subscribeTimestamp[subLoopVariable].tv_nsec,
