+++ examples/pubsub_realtime/pubsub_TSN_publisher.c
@@ -78,7 +78,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <linux/types.h>
-#include <sys/io.h>
+//#include <sys/io.h>
 #include <getopt.h>
 
 /* For thread operations */
@@ -350,7 +350,7 @@ addPubSubApplicationCallback(UA_Server *server, UA_NodeId identifier,
         *callbackId = threadCreation((UA_Int16)pubPriority, (size_t)pubCore,
                                      publisherETF, threadNamePub, threadArguments);
         UA_LOG_INFO(UA_Log_Stdout, UA_LOGCATEGORY_USERLAND,
-                    "Publisher thread callback Id: %ld\n", *callbackId);
+                    "Publisher thread callback Id: %lld\n", *callbackId);
 #endif
     }
     else {
@@ -360,7 +360,7 @@ addPubSubApplicationCallback(UA_Server *server, UA_NodeId identifier,
         *callbackId = threadCreation((UA_Int16)subPriority, (size_t)subCore,
                                      subscriber, threadNameSub, threadArguments);
         UA_LOG_INFO(UA_Log_Stdout, UA_LOGCATEGORY_USERLAND,
-                    "Subscriber thread callback Id: %ld\n", *callbackId);
+                    "Subscriber thread callback Id: %lld\n", *callbackId);
 #endif
     }
 
@@ -381,9 +381,9 @@ changePubSubApplicationCallback(UA_Server *server, UA_NodeId identifier,
 static void
 removePubSubApplicationCallback(UA_Server *server, UA_NodeId identifier,
                                 UA_UInt64 callbackId) {
-    if(callbackId && (pthread_join(callbackId, NULL) != 0))
+    if(callbackId && (pthread_join((long unsigned int)callbackId, NULL) != 0))
         UA_LOG_WARNING(UA_Log_Stdout, UA_LOGCATEGORY_USERLAND,
-                       "Pthread Join Failed thread: %ld\n", callbackId);
+                       "Pthread Join Failed thread: %lld\n", callbackId);
 }
 
 /**
@@ -902,7 +902,7 @@ updateMeasurementsPublisher(struct timespec start_time,
 
     if(consolePrint)
         UA_LOG_INFO(UA_Log_Stdout, UA_LOGCATEGORY_USERLAND,
-                    "Pub:%ld,%ld.%09ld\n", counterValue, start_time.tv_sec, start_time.tv_nsec);
+                    "Pub:%lld,%ld.%09ld\n", counterValue, start_time.tv_sec, start_time.tv_nsec);
 
     if(signalTerm != true){
         publishTimestamp[measurementsPublisher]        = start_time;
@@ -931,7 +931,7 @@ updateMeasurementsSubscriber(struct timespec receive_time,
     }
 
     if(consolePrint)
-        UA_LOG_INFO(UA_Log_Stdout, UA_LOGCATEGORY_USERLAND, "Sub:%ld,%ld.%09ld\n",
+        UA_LOG_INFO(UA_Log_Stdout, UA_LOGCATEGORY_USERLAND, "Sub:%lld,%ld.%09ld\n",
                     counterValue, receive_time.tv_sec, receive_time.tv_nsec);
 
     if(signalTerm != true)
@@ -1219,7 +1219,7 @@ threadCreation(UA_Int16 threadPriority, size_t coreAffinity,
 
     if(CPU_ISSET(coreAffinity, &cpuset))
         UA_LOG_INFO(UA_Log_Stdout, UA_LOGCATEGORY_USERLAND,
-                    "%s CPU CORE: %ld\n", applicationName, coreAffinity);
+                    "%s CPU CORE: %d\n", applicationName, coreAffinity);
 
    return threadID;
 }
@@ -1417,7 +1417,7 @@ static void computeLatencyAndGenerateCsv(char *latencyFileName) {
 
         if(((latencyCharIndex - prevLatencyCharIndex) + latencyCharIndex + 3) < MAX_MEASUREMENTS_FILEWRITE) {
             latencyCharIndex += (UA_UInt64)sprintf(&latency_measurements[latencyCharIndex],
-                                                   "%0.3f, %ld, %ld\n",
+                                                   "%0.3f, %lld, %lld\n",
                                                    finalTime, missed_counter, repeated_counter);
         }
         else {
@@ -1430,7 +1430,7 @@ static void computeLatencyAndGenerateCsv(char *latencyFileName) {
     }
 
     /* Write into the latency file */
-    fwrite(&latency_measurements[0], prevLatencyCharIndex, 1, fp_latency);
+    fwrite(&latency_measurements[0], (unsigned int)prevLatencyCharIndex, 1, fp_latency);
     fclose(fp_latency);
 }
 
@@ -1719,7 +1719,7 @@ int main(int argc, char **argv) {
         size_t pubLoopVariable               = 0;
         for(pubLoopVariable = 0; pubLoopVariable < measurementsPublisher;
              pubLoopVariable++) {
-            fprintf(fpPublisher, "%ld,%ld.%09ld\n",
+            fprintf(fpPublisher, "%lld,%ld.%09ld\n",
                     publishCounterValue[pubLoopVariable],
                     publishTimestamp[pubLoopVariable].tv_sec,
                     publishTimestamp[pubLoopVariable].tv_nsec);
@@ -1730,7 +1730,7 @@ int main(int argc, char **argv) {
         size_t subLoopVariable               = 0;
         for(subLoopVariable = 0; subLoopVariable < measurementsSubscriber;
              subLoopVariable++) {
-            fprintf(fpSubscriber, "%ld,%ld.%09ld\n",
+            fprintf(fpSubscriber, "%lld,%ld.%09ld\n",
                     subscribeCounterValue[subLoopVariable],
                     subscribeTimestamp[subLoopVariable].tv_sec,
                     subscribeTimestamp[subLoopVariable].tv_nsec);
